Fixed #381: Change property ByHandleFileInfo.VolumeSerialNumber from int to long.

+++ AlphaFS/Filesystem/ByHandleFileInfo.cs
@@ -40,7 +40,7 @@ namespace Alphaleonis.Win32.Filesystem
          FileIndex = NativeMethods.ToLong(fibh.nFileIndexHigh, fibh.nFileIndexLow);
          FileSize = NativeMethods.ToLong(fibh.nFileSizeHigh, fibh.nFileSizeLow);
          NumberOfLinks = (int) fibh.nNumberOfLinks;
-         VolumeSerialNumber = (int) fibh.dwVolumeSerialNumber;
+         VolumeSerialNumber = fibh.dwVolumeSerialNumber;
       }
 
 
@@ -108,7 +108,7 @@ namespace Alphaleonis.Win32.Filesystem
 
       /// <summary>Gets the serial number of the volume that contains a file.</summary>
       /// <value>The serial number of the volume that contains a file.</value>
-      public int VolumeSerialNumber { get; private set; }
+      public long VolumeSerialNumber { get; private set; }
 
 
       /// <summary>Gets the size of the file.</summary>
